Implement top level song menu
