+++ Persephone/Controllers/AlbumDetailView.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2019 Dan Barber. All rights reserved.
 //
 
-import Cocoa
+import AppKit
 
 class AlbumDetailView: NSViewController {
   var album: Album?
@@ -60,6 +60,8 @@ class AlbumDetailView: NSViewController {
     albumTitle.stringValue = ""
     albumArtist.stringValue = ""
     albumCoverView.image = .defaultCoverArt
+
+    App.store.dispatch(SetSelectedSong(selectedSong: nil))
   }
 
   @IBAction func playAlbum(_ sender: NSButton) {
@@ -138,85 +140,5 @@ class AlbumDetailView: NSViewController {
 
   func setAlbum(_ album: Album) {
     self.album = album
-  }
-}
-
-extension AlbumDetailView: NSTableViewDelegate {
-  func tableView(_ tableView: NSTableView, viewFor tableColumn: NSTableColumn?, row: Int) -> NSView? {
-    if let song = dataSource.albumSongs[row].song {
-      switch tableColumn?.identifier.rawValue {
-      case "trackNumberColumn":
-        return cellForTrackNumber(tableView, with: song)
-      case "trackTitleColumn":
-        return cellForSongTitle(tableView, with: song)
-      case "trackDurationColumn":
-        return cellForSongDuration(tableView, with: song)
-      default:
-        return nil
-      }
-    } else if let disc = dataSource.albumSongs[row].disc {
-      return cellForDiscNumber(tableView, with: disc)
-    }
-
-    return nil
-  }
-
-  func tableView(_ tableView: NSTableView, isGroupRow row: Int) -> Bool {
-    return dataSource.albumSongs[row].disc != nil
-  }
-
-  func tableView(_ tableView: NSTableView, rowViewForRow row: Int) -> NSTableRowView? {
-    let view = AlbumDetailSongRowView()
-
-    return view
-  }
-
-  func tableView(_ tableView: NSTableView, shouldSelectRow row: Int) -> Bool {
-    return dataSource.albumSongs[row].disc == nil
-  }
-
-  func cellForDiscNumber(_ tableView: NSTableView, with disc: String) -> NSView {
-    let cellView = tableView.makeView(
-      withIdentifier: .discNumber,
-      owner: self
-      ) as! NSTableCellView
-
-    cellView.textField?.stringValue = "Disc \(disc)"
-
-    return cellView
-  }
-
-  func cellForTrackNumber(_ tableView: NSTableView, with song: Song) -> NSView {
-    let cellView = tableView.makeView(
-      withIdentifier: .trackNumber,
-      owner: self
-      ) as! NSTableCellView
-
-    cellView.textField?.stringValue = "\(song.trackNumber)."
-
-    return cellView
-  }
-
-  func cellForSongTitle(_ tableView: NSTableView, with song: Song) -> NSView {
-    let cellView = tableView.makeView(
-      withIdentifier: .songTitle,
-      owner: self
-    ) as! NSTableCellView
-
-    cellView.textField?.stringValue = song.title
-
-    return cellView
-  }
-
-  func cellForSongDuration(_ tableView: NSTableView, with song: Song) -> NSView {
-    let cellView = tableView.makeView(
-      withIdentifier: .songDuration,
-      owner: self
-      ) as! NSTableCellView
-
-    cellView.textField?.font = .timerFont
-    cellView.textField?.stringValue = song.duration.formattedTime
-
-    return cellView
   }
 }
